libwasi: move proc and sched to separate files
